Check 'after signed out' case & some refactoring
Adds new spec to cover the case where a user has an account, but has
signed out of it.

This commit uses shared examples and some helper methods to make the
acceptance specs read more nicely (and make sure their brittleness
isn't duplicated across multiple places, yay)

This commit also fixes a bug with using ActionMailer in tests,
clearing out the state between tests in the spec helper.
