+++ spec/features/asking_users_to_sign_in_spec.rb
@@ -1,7 +1,7 @@
-describe 'asking users to sign in' do
-  context 'when I try to visit the homepage' do
-    before { visit '/' }
+require_relative 'support/sign_up_helpers'
 
+describe 'asking users to sign in' do
+  shared_examples 'I am not signed in' do
     it 'asks for an email' do
       expect(page).to have_field 'Email'
     end
@@ -13,5 +13,26 @@ describe 'asking users to sign in' do
     it 'asks me to sign in' do
       expect(page).to have_button 'Log in'
     end
+  end
+
+  context 'with no account' do
+    context 'when I try to visit the homepage' do
+      before { visit '/' }
+
+      it_behaves_like 'I am not signed in'
+    end
+  end
+
+  context 'with a pre-existing account, but signed out' do
+    before do
+      sign_up_for_account
+      create_password_for_account
+
+      click_on 'Logout'
+
+      visit '/'
+    end
+
+    it_behaves_like 'I am not signed in'
   end
 end
