+++ spec/features/support/sign_up_helpers.rb
@@ -1,0 +1,18 @@
+def sign_up_for_account
+  visit new_user_registration_path
+  fill_in 'user_email', with: 'tom@tom.com'
+  click_on 'Sign up'
+end
+
+def create_password_for_account
+  visit confirmation_email_link
+  fill_in 'New password', with: 'supersecret'
+  fill_in 'Confirm new password', with: 'supersecret'
+  click_on 'Change my password'
+end
+
+def confirmation_email_link
+  confirmation_email = ActionMailer::Base.deliveries.first
+  parsed_email = Nokogiri::HTML(confirmation_email.body.to_s)
+  parsed_email.css('a').first['href']
+end
