+++ spec/spec_helper.rb
@@ -8,4 +8,6 @@ RSpec.configure do |config|
   end
 
   config.shared_context_metadata_behavior = :apply_to_host_groups
+
+  config.before { ActionMailer::Base.deliveries.clear }
 end
